+++ src/texto/extensions/italic/italic.ts
@@ -47,7 +47,6 @@ export const Italic = Mark.create<ItalicOptions>({
   addKeyboardShortcuts() {
     return {
       'Mod-i': () => this.editor.commands.toggleItalic(),
-      'Mod-I': () => this.editor.commands.toggleItalic(),
     }
   },