+++ src/texto/extensions/underline/underline.ts
@@ -37,7 +37,6 @@ export const Underline = Mark.create<UnderlineOptions>({
   addKeyboardShortcuts() {
     return {
       'Mod-u': () => this.editor.commands.toggleUnderline(),
-      'Mod-U': () => this.editor.commands.toggleUnderline(),
     }
   },
 })